refactor(LayoutSetter): 移除 LayoutModal 组件，整合布局设置逻辑，优化样式管理
