+++ src/editor/components/Setting/ComponentStyle/LayoutSetter/index.tsx
@@ -10,9 +10,12 @@
 import type { CSSProperties } from "react";
 import type { Component } from "../../../../stores/components";
 import BoxModelEditor from "./BoxModelEditor";
-import LayoutModal from "./LayoutModal";
 import SizeEditor from "./SizeEditor";
 import { Form } from "antd";
+import FormItem from "antd/es/form/FormItem";
+import StyleOptionGroup from "../../../common/StyleOptionGroup";
+import { useStyleChangeHandler } from "../../../../hooks/useStyleChangeHandler";
+import { layoutSettings } from "./config";
 
 interface LayoutProps {
   curComponent: Component;
@@ -21,6 +24,9 @@ interface LayoutProps {
 
 const LayoutSetter = (props: LayoutProps) => {
   const { curComponent, onChange } = props;
+  const createChangeHandler = useStyleChangeHandler(onChange);
+  const value = curComponent.styles || {};
+  const displayValue = value.display;
 
   return (
     <Form
@@ -29,17 +35,52 @@ const LayoutSetter = (props: LayoutProps) => {
       labelAlign="right"
       style={{ width: "100%" }}
     >
-      <LayoutModal value={curComponent.styles} onChange={onChange} />
-      <Form.Item label="边距">
+      <StyleOptionGroup
+        label="布局模式"
+        options={layoutSettings.display}
+        currentValue={displayValue}
+        onChange={createChangeHandler("display")}
+      />
+
+      {/* 条件渲染 Flex 布局的专属设置 */}
+      {displayValue === "flex" && (
+        <div className="pl-9 border-l-2 border-gray-100 ml-3">
+          <StyleOptionGroup
+            label="主轴方向"
+            options={layoutSettings.flexDirection}
+            currentValue={value.flexDirection}
+            onChange={createChangeHandler("flexDirection")}
+          />
+          <StyleOptionGroup
+            label="主轴对齐"
+            options={layoutSettings.justifyContent}
+            currentValue={value.justifyContent}
+            onChange={createChangeHandler("justifyContent")}
+          />
+          <StyleOptionGroup
+            label="辅轴对齐"
+            options={layoutSettings.alignItems}
+            currentValue={value.alignItems}
+            onChange={createChangeHandler("alignItems")}
+          />
+          <StyleOptionGroup
+            label={"换\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0行"} // 使用 Unicode 空格确保对齐
+            options={layoutSettings.flexWrap}
+            currentValue={value.flexWrap}
+            onChange={createChangeHandler("flexWrap")}
+          />
+        </div>
+      )}
+      <FormItem label="边距">
         <BoxModelEditor value={curComponent.styles} onChange={onChange} />
-      </Form.Item>
-      <Form.Item label="大小">
+      </FormItem>
+      <FormItem label="大小">
         <SizeEditor
           value={curComponent.styles}
           onChange={onChange}
           curComponentId={curComponent.id}
         />
-      </Form.Item>
+      </FormItem>
     </Form>
   );
 };
